Refactor ascending/ descending iteration approach
The point of all this refactoring was to get the body of for loops in the coloring methods in the different colorer classes (SequentialGreedy and LubyJonesGreedy) to be their own functions. This is because iterating forwards or backwards (which is needed for the ascending boolean parameter) requires using different iterators (begin vs rbegin to end vs rend).

So the solution was to make the classes non-static and add constructors that take info about the graph (namely, the adjacency list and the sorted vertices). Then the color() functions only take an ascending variable, and further reference the private variables that were initialized in the constructors.

Depending on the direction (ascending/ descending), the color() functions use a forwards or backwards iterator in a one-line for loop that just calls the function responsible for doing the actual work of a single iteration. This function is called with references to some of the variables created by the color function, and is also able to access the information about the graph which was initialized in the constructor.

So now finally the insides of the for loops that did most the work doesn't have to be duplicated in the code anymore.

This also made it easier for all the colorer classes to have a single Colorer class as base, which has variables they all need (such as the ones keeping track of the graph) along with methods they all need (such as finding the lowest allowed color for a vertex).

Documentation is also improved.