Refactor from color map to colors in main vector
The previous refactor enabled each remapped vertex (indexed 0 to N-1) to store both their color and neighbors. This made the unsorted ColorMap previously used to map vertices to their colors unnecessary; this refactor enables using the color location built in to the main vector that stores the vertices.

Around a 3x speed up!